o1vm/mips: move tests reg. total nb of constraints and max degree
