+++ o1vm/src/pickles/mod.rs
@@ -17,21 +17,12 @@ pub mod proof;
 pub mod prover;
 pub mod verifier;
 
-/// Maximum degree of the constraints.
-/// It does include the additional degree induced by the multiplication of the
-/// selectors.
-pub const MAXIMUM_DEGREE_CONSTRAINTS: u64 = 6;
-
 /// Degree of the quotient polynomial. We do evaluate all polynomials on d8
 /// (because of the value of [MAXIMUM_DEGREE_CONSTRAINTS]), and therefore, we do
 /// have a degree 7 for the quotient polynomial.
 /// Used to keep track of the number of chunks we do have when we commit to the
 /// quotient polynomial.
 pub const DEGREE_QUOTIENT_POLYNOMIAL: u64 = 7;
-
-/// Total number of constraints for all instructions, including the constraints
-/// added for the selectors.
-pub const TOTAL_NUMBER_OF_CONSTRAINTS: usize = 466;
 
 #[cfg(test)]
 mod tests;
